CRO-Backend/CroBackend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
altgraph==0.17.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastuuid==0.13.5
filelock==3.19.1
frozenlist==1.6.0
fsspec==2025.9.0
future==0.18.2
griffe==1.14.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.77.7
macholib==1.15.2
MarkupSafe==3.0.3
multidict==6.4.4
openai==1.109.1
openai-agents==0.3.3
packaging==25.0
pillow==11.3.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.3
rpds-py==0.27.1
six==1.15.0
sniffio==1.3.1
sqlparse==0.5.3
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
yarl==1.20.0
zipp==3.23.0
gunicorn==21.2.0
